requirements.txt
streamlit
python-dotenv
groq
requests
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers